Fix for hang in Ubuntu 15.04
Problem: The thread safe queue was being destroyed before the threads
that were reading from it. This caused a hang, as the threads were
continuously reading and popping, but the queue was dead so the threads
never woke up. Now we sleep for 100 ms instead. This is not an ideal
fix, but it causes all tests to pass, showing that this at least in
principle solves the problem.