Correction of internal fs given ppm correction
This corrects the internal_fs_hz configuration parameter if a
SignalSource.clock_correction_ppm parameter is specified.

This allows us to specify our desired sampling rate as say 2 MHz, and
automatically correcting the sampling rate used in the receiver to

2MHz*( 1 - clock_correction_ppm/1e6 )

Useful for real time sources whose clocks are not the most accurate
(e.g. the HackRF has a clock uncertainty of ~ 10ppm )